My 6-size commit

+++ 0x00-hello_world/6-size.c
@@ -10,10 +10,10 @@ int n;
 long int o;
 long long int p;
 float q;
-printf("Size of a char: %lu byte(s)\n", sizeof(m));
-printf("Size of int: %lu byte(s)\n", sizeof(n));
-printf("Size of long int: %lu byte(s)\n", sizeof(o));
-printf("Size of long long int: %lu byte(s)\n", sizeof(p));
-printf("Size of float: %lu byte(s)\n", sizeof(q));
+printf("Size of a char: %lu byte(s)\n", (unsigned long) sizeof(m));
+printf("Size of int: %lu byte(s)\n", (unsigned long)sizeof(n));
+printf("Size of long int: %lu byte(s)\n", (unsigned long)sizeof(o));
+printf("Size of long long int: %lu byte(s)\n", (unsigned long)sizeof(p));
+printf("Size of float: %lu byte(s)\n", (unsigned long)sizeof(q));
 return (0);
 }
